Still working on ThePurge...

+++ ThePurge/Main.java
@@ -13,6 +13,8 @@ public class Main
     static ArrayList<String> in = new ArrayList<String>();
 
     static ArrayList<ArrayList<String[]>> folders = new ArrayList<ArrayList<String[]>>();
+    static int numFoldersDeleted = 0;
+
 
     public static void main(String[] args)
     {
@@ -41,30 +43,39 @@ public class Main
 
         ArrayList<String[]> currFolder;
         String[] currFile;
-        int sumOfBytesDeleted = 0;
+        int sumOfBytesDeleted = 0, currSize = 0;
 
         for(int i = 0; i < folders.size(); i++){
             currFolder = folders.get(i);
-            System.out.println("Going through folder " + i + " now with " + currFolder.size() + " items");
+            
+            if(currFolder != null){
+                currSize = currFolder.size();
+            } else {
+                currSize = 0;
+            }
 
-            for(int j = 0; j < currFolder.size(); j++){
-                System.out.println("running...");
-                currFile = currFolder.get(j);
+            System.out.println("Going through folder " + i + " now with " + currSize + " items");
 
-                if(currFile[0].contains("delete") || currFile[0].contains("temporary")){
-                    //check if it is a file or a folder
-                    System.out.println("Processing..." + currFile[0] + " with file info " + currFile[1]);
-                    if(currFile[1].indexOf("FOLDER") != -1){
-                        int folderNum = Integer.parseInt(currFile[1].substring(8, 9));
-                        System.out.println("Getting ready to delete Folder" + folderNum);
-                        sumOfBytesDeleted+= removeDirectory(folderNum);
-                    } else {
-                        sumOfBytesDeleted+= Integer.parseInt(currFile[1]);
-                        currFolder.remove(j);
-                        j--;
+            if(currFolder != null){
+                for(int j = 0; j < currFolder.size(); j++){
+                    System.out.println("running...");
+                    currFile = currFolder.get(j);
+
+                    if(currFile[0].contains("delete") || currFile[0].contains("temporary")){
+                        //check if it is a file or a folder
+                        System.out.println("Processing..." + currFile[0] + " with file info " + currFile[1]);
+                        if(currFile[1].indexOf("FOLDER") != -1){
+                            int folderNum = Integer.parseInt(currFile[1].substring(8, 9));
+                            System.out.println("Getting ready to delete Folder " + folderNum);
+                            sumOfBytesDeleted+= removeDirectory(folderNum);
+                        } else {
+                            sumOfBytesDeleted+= Integer.parseInt(currFile[1]);
+                            currFolder.remove(j);
+                            j--;
+                        }
                     }
-                }
 
+                }
             }
 
         }
@@ -124,27 +135,36 @@ public class Main
         String[] currFile;
         int sumOfBytesDeleted = 0;
 
-        //Go through everything in the folder and add the bytes
-        for(int j = 0; j < currFolder.size(); j++){
-            currFile = currFolder.get(0);
+        if(currFolder != null){
+            //Go through everything in the folder and add the bytes
+            for(int j = 0; j < currFolder.size(); j++){
+                currFile = currFolder.get(0);
 
 
-            //check if it is a file or a folder
-            if(currFile[1].contains("FOLDER")){
-                int folderNum = Integer.parseInt(currFile[1].substring(8, 9));
-                System.out.println("\tGetting ready to delete Folder " + folderNum);
-                sumOfBytesDeleted+= removeDirectory(folderNum);
-            } else {
-                sumOfBytesDeleted+= Integer.parseInt(currFile[1]);
-                currFolder.remove(j);
-                j--;
+                //check if it is a file or a folder
+                if(currFile != null){
+                    if (currFile[1].contains("FOLDER")){
+                        int folderNum = Integer.parseInt(currFile[1].substring(8, 9));
+                        System.out.println("\tGetting ready to delete Folder " + folderNum);
+                        sumOfBytesDeleted+= removeDirectory(folderNum);
+                    } else {
+                        sumOfBytesDeleted+= Integer.parseInt(currFile[1]);
+                        currFolder.remove(j);
+                        j--;
+                    }
+
+                    
+                }
             }
-            
+            System.out.println("removed Folder " + dir + " saved " + sumOfBytesDeleted + " bytes");
+            folders.set(dir, null);
         }
 
-        System.out.println("removed Folder " + dir);
+        System.out.println("Folder " + dir + " is empty.");
 
-        folders.remove(dir);
+
+
+        //folders.remove(dir);
         return sumOfBytesDeleted;
     }
